Removed unused files and fixed linting
