+++ workshops/introduction_to_event_sourcing/src/solved/05_business_logic/immutable/solution1/businessLogic.ts
@@ -4,7 +4,6 @@
 
 import {
   EventStore,
-  evolve,
   PricedProductItem,
   ProductItemAddedToShoppingCart,
   ProductItemRemovedFromShoppingCart,
